Fix crash in rain strobe pattern
This pattern assumes that only white pixels are present when the pattern
starts, which should be true if other patterns have correctly
transitioned out.  But if it isn't true the pattern crashed.  Bodge this
issue by forcing any non-white pixels to be white at the start of this
pattern instead of by crashing.  This issue happens so infrequently that
the slightly inelegance is not worth reworking this pattern's code.